Migrate utilities to main xiaomi file
